+++ test/FunctionMetadataGeneratorTests/FunctionMetadataGeneratorTests.cs
@@ -212,10 +212,8 @@ namespace Microsoft.Azure.Functions.SdkTests
                 {
                     { "Name", "blobinput" },
                     { "Type", "blob" },
-                    { "DataType", "String"},
                     { "Direction", "In" },
                     { "blobPath", "container2" },
-                    { "Cardinality", "Many" },
                     { "Properties", new Dictionary<String, Object>( ) { { "SupportsDeferredBinding" , "True"} } }
                 });
             }
@@ -325,7 +323,6 @@ namespace Microsoft.Azure.Functions.SdkTests
                     { "Type", "blob" },
                     { "Direction", "In" },
                     { "blobPath", "container2/%file%" },
-                    { "Cardinality", "One" },
                     { "Properties", new Dictionary<String, Object>( ) { { "SupportsDeferredBinding" , "True"} } }
                 });
             }
@@ -375,7 +372,6 @@ namespace Microsoft.Azure.Functions.SdkTests
                     { "Type", "blob" },
                     { "Direction", "In" },
                     { "blobPath", "container2" },
-                    { "Cardinality", "Many" },
                     { "Properties", new Dictionary<String, Object>( ) { { "SupportsDeferredBinding", "True" } } }
                 });
             }
@@ -395,8 +391,6 @@ namespace Microsoft.Azure.Functions.SdkTests
                     { "Type", "blob" },
                     { "Direction", "In" },
                     { "blobPath", "container2" },
-                    { "Cardinality", "Many" },
-                    { "DataType", "String" },
                     { "Properties", new Dictionary<String, Object>( ) { { "SupportsDeferredBinding", "True" } } }
                 });
             }
@@ -1169,7 +1163,7 @@ namespace Microsoft.Azure.Functions.SdkTests
             [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
             public object BlobToBlobs(
                 [BlobTrigger("container2/%file%")] string blob,
-                [BlobInput("container2", IsBatched = true)] IEnumerable<string> blobinput)
+                [BlobInput("container2")] IEnumerable<string> blobinput)
             {
                 throw new NotImplementedException();
             }
@@ -1269,7 +1263,7 @@ namespace Microsoft.Azure.Functions.SdkTests
             [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
             public object BlobStringToBlobStringArray(
                 [BlobTrigger("container2/%file%")] string blob,
-                [BlobInput("container2", IsBatched = true)] string[] blobinput)
+                [BlobInput("container2")] string[] blobinput)
             {
                 throw new NotImplementedException();
             }
@@ -1279,7 +1273,7 @@ namespace Microsoft.Azure.Functions.SdkTests
             [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
             public object BlobStringToBlobClientEnumerable(
                 [BlobTrigger("container2/%file%")] string blob,
-                [BlobInput("container2", IsBatched = true)] IEnumerable<BlobClient> blobinput)
+                [BlobInput("container2")] IEnumerable<BlobClient> blobinput)
             {
                 throw new NotImplementedException();
             }
@@ -1288,7 +1282,7 @@ namespace Microsoft.Azure.Functions.SdkTests
             [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
             public object BlobStringToBlobPocoEnumerable(
                 [BlobTrigger("container2/%file%")] string blob,
-                [BlobInput("container2", IsBatched = true)] IEnumerable<Poco> blobinput)
+                [BlobInput("container2")] IEnumerable<Poco> blobinput)
             {
                 throw new NotImplementedException();
             }
@@ -1297,7 +1291,7 @@ namespace Microsoft.Azure.Functions.SdkTests
             [BlobOutput("container1/hello.txt", Connection = "MyOtherConnection")]
             public object BlobStringToBlobPocoArray(
                 [BlobTrigger("container2/%file%")] string blob,
-                [BlobInput("container2", IsBatched = true)] Poco[] blobinput)
+                [BlobInput("container2")] Poco[] blobinput)
             {
                 throw new NotImplementedException();
             }