+++ extensions/Worker.Extensions.Storage.Blobs/release_notes.md
@@ -4,6 +4,7 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Extensions.Storage.Blobs <version>
+### Microsoft.Azure.Functions.Worker.Extensions.Storage.Blobs 6.0.0
 
-- <entry>
+- Remove IsBatched property from BlobInput binding attribute
+- Infer binding cardinality based on blobPath